@dscho 2024-01-11
> The pull request has 46 commits. The max allowed is 30. Please split the patch series into multiple pull requests. Also consider squashing related commits.

If you plan on submitting this via GitGitGadget, we'll need an exception similar to https://github.com/gitgitgadget/gitgitgadget/pull/619.