@pks-t 2024-01-11
> If you plan on submitting this via GitGitGadget, we'll need an exception similar to https://github.com/gitgitgadget/gitgitgadget/pull/619.

I don't, but thanks for checking in! For one almost all of the commits will already land via already-in-flight topics. But second I merely wanted to check whether the reftable backend will hit GitHub Actions specific issues (which it did).

Two of the issues reported are known already and already fixed. But the Perforce issue is new to me. Should be an easy fix though, I'll send a patch upstream soonish.