@dscho 2024-01-11
> I merely wanted to check whether the reftable backend will hit GitHub Actions specific issues (which it did).

Excellent!